Fixed Grain and Food Class
In Food class:
I wasn;t properly inheriting the super(), but now I am using the arguments(maturation_age,death_age, monetary_value) which were not there before.

And in Grain Class
made the generationcount varaible static(was not static earlier). I tried  making the object but the counter was not incrementing but using static makes it possible to increase it as earlier it was making count for eah individual seperate, but now increasing based on the grain objects made. I did testing of it in make file.