Fixed functionality of a method in Apple class and Item class
applied the same thing which i did for counting the generationcount, added increement statement just after te super constructor so that it can keep track of the objects being created every time.

In Item class:
Fixed died method:
As it was returning true regardless of the case provided and was also checking if the age is greater than equal to but in assignment scenario they asked for just greater than hence i removed the "=" operator. And tried running it without if +-else, and turns-out it works, so keeping it.

And for getValue() method i changed the return to 0, because when i ran it prints 3, but in the marker.java fine on line no:20 it says it should print 0. so i changed that to 0. Not sure if that's the correct way.