Added tick() un field class
First, i Imported soil,weed,untilledsoil class along with the item class.
then made height and width variable private as before they were not, made a private 2d array that takes item objects inside it, and a random number gernerator for the 20% of chance of growing weed on the field.

In Field constructor, i only pass height and width as a parameter since those are the one we need as a parameter to make a field according to our liking, field which takes height and width variable as the actual dimensions of the field.
the nested for loop job is to put soil on every columns(width) and rows(height) of the field.

On to the main functionality of tick.
used the same nested for as in constructor, but this time it will call the tick() method on itself to increase the age of every single soil object we are looping through.

The nested if is responsible for checking if it is a soil, if yes then
if random number is 1, make the block(field[i][j]) weed.

The last if checks if the current item is died (checks if the item's age is past its death age from Item class died method)
If it is , then make the item blockin the field (on which we are on i.e: field[i][j]) to untilledsoil.