Added plant() and get() method in field class
Plant method which takes row, colum and item(Item object for planting weed, untilledsoil, grain apple) as a parameter. It plants the item on to the current location in the field(which is field[row][column]).

get() method takes row and column as a parameter(as mentioned). Made a copy of the current location and stored it in originalcopy(apologise for this name). Then checks if the copy item is instance of weed or soil, if it is, then sets the their age by calling getAge from copy variable made above and returns the copy of it. However, I am not sure if I am doing it right or wrong,  also it justs says on the assignment that it basically returns but there's lot of case to deal with. Like i am only checking if the copy is soil or weed only, nit much. Leaving this as it is. Don't know whether i am right or wrong.