Added getValue() in Field class.
responsible for getting total monetary value of each item.

Made a counter that stores and adds the monetary value, and then returning it,
I test it in the marker class, by making an apple and a grain object, since apple mature age is 3, so i ran tick() (which matures it) three times. Just to be sure the value was changing, i print the value of it, from both testfield and apple class.