Fixed issues in run()
moved the scanner object which was in loop, to outside the loop, aling with other variables which are declared in the beginning, fixed the handilin og while loop. was not showing anything to the console before, as i think because it was infinite, and also the naming convention was a bit confusing. kept the name to run now and made it a boolean, to use that to make while loop run correctly.