Fixed quit method and if-else statements in run()
First-  Fix the if-else, as == was checking just the refernce not the actual string, that;s why use .equals().

Second- quit() logic was not working properly, so i made the run(which was local to run() method) global. Now i can use that to change it to false in quit() method, to make it work properly.