removed the plant method
Removed the plant method, since i was having issues running it. Running it from run() method(As i moved and added the logic inside run()). using an array to store the string(which are splitted using .split), and converting those string to integers. storing the field spot(spot object in the field like soil or untilled soil) in item, and checking if it is soil. prompts the user to buy either apple or grain but incomplete yet .still need to complete the below part.