Fixing and adding onto plant logic
Added the cost of buying an apple and grain in the user prompt section.

And now using choice which is local instead of action(which is global and also it changes the first prompt entered by the user.

Added if logic to check if user entered a, deduct the money from balance, and plant the apple. Same logic for grain. In else if part, checks if the user has less balance then show the message.