Fixed the current if-else plant logic
removed unnecessary else if as they are causing error, since the else itself covers all the cases. imported apple and grain class, and used them as new objects in planting, since item refers to the soil, and it does not make sense to plant soil on top of soil. Also changed the print statements in else block for balance.
And added a seprate that handle the case where user enters something else.
First is if user enters anything instead of 'a' and 'g'. for eg: user enters w 2 5, it will display the following message.
'Second else to check if the item isnot soil, show the message.