Added Harvest
Since it took me nearly three days to figure the plant logic, harvest and till logic is kind of same with differnce in action.

Harvest here does the same as checking the length of string entered by user and if the first letter is 'h', convert the second and third numbers(which are string) to int, and stores the copy of the item location in item.
if checks if it is food and its maturationage is less than the curent age. increases balance and tills the soil