Added the getter for maturation age
Forgot to mention, should have done before the last comment. Since in harvest it requires maturation age. I haven't added the maturation age getter to access its mature age. but now I added it so that it can be used in the if block of harvest to compare it to items current age.